Group library macros by category in a tree view, fix line width doubling everywhere
The macro library panel used a flat icon list with a plain (non-selectable)
text item as a fake category separator. Replaced it with a QTreeWidget per
library page - category as parent node, macros as icon+name leaves - so the
.fcl format's own category grouping actually reads as sections instead of
same-sized cells in a grid. Search filtering now hides a category node too
once none of its macros match, instead of always leaving it visible.

Separately: every schematic primitive (Line, Bezier, Rectangle, Ellipse,
Polygon, ComplexCurve) was hardcoded to a pen width of 1 unit, and the field
storing it (GraphicsPrimitive::penSize) was an int - silently unable to
represent the reference FidoCadJ editor's actual default of 0.5
(Globals.lineWidth, confirmed straight from PrimitiveLine.java/PrimitiveOval.
java/etc., all of which stroke at that same shared value rather than carrying
a per-primitive width token). That made every one of these primitives twice
as thick as it should be - most visible in densely-packed library macros,
which is what made it look like a macro-specific bug rather than a global
one. Widened penSize to qreal and wired both FidoCadReader (reading a file or
a macro body) and PrimitivePlacementController (interactive drawing) to set
it from the document's actual line width (Sheet::lineWidth(), already
correctly defaulted to 0.5 and already FJC-"A"-overridable, just never
applied to anything until now). PL/PA (PCB track/pad) already carry their own
explicit width token and are unaffected.

+++ src/GraphicsPrimitive.h
@@ -75,7 +75,7 @@ public:
     // these each call) - needed by FidoCadWriter, which isn't a subclass and
     // so can't reach the protected fields directly.
     Qt::PenStyle lineStyle() const { return penStyle; }
-    int lineWidth() const { return penSize; }
+    qreal lineWidth() const { return penSize; }
 
     void setName(const QString &name) { objName = name; }
     void setValue(const QString &value) { objValue = value; }
@@ -184,7 +184,7 @@ signals:
 public slots:
 
     void penStyleIsChanged(Qt::PenStyle penStyle) { this->penStyle = penStyle; }
-    void setPenSize(int newPenSize) { penSize = newPenSize; }
+    void setPenSize(qreal newPenSize) { penSize = newPenSize; }
     void setIsFilled(bool isFilled) { filled = isFilled; }
     void setNameVisible(bool visible) { showName = visible; }
     void setValueVisible(bool visible) { showValue = visible; }
@@ -209,7 +209,7 @@ protected:
     QString  objValue;
     Layer *objLayer;
     QPen _pen;
-    int penSize;
+    qreal penSize;
 
     QPointF m_dragAnchor; // last mouse scene position seen during an active drag
     // Control-point snapshots (per moved primitive - a multi-selection drag
